added host support
