+++ showcase-api/envs/dev.yaml
@@ -1,2 +1,3 @@
 server-config:
+  host: 'localhost'
   port: 9090
