+++ showcase-api/envs/prod.yaml
@@ -1,2 +1,3 @@
 server-config:
+  host: 'example.com'
   port: 80
