+++ showcase-api/envs/sandbox.yaml
@@ -1,2 +1,3 @@
 server-config:
+  host: 'https://example.com'
   port: 81
